refactor: move SqlitePool to authly-sqlite
