+++ lib/authly-sqlite/src/param.rs
@@ -1,12 +1,6 @@
-//! sqlite integration with the `Db` trait
-//!
-//! This Db implementation is used in tests for now.
-
 use core::str;
 
 use rusqlite::{params_from_iter, types::ToSqlOutput, ParamsFromIter};
-
-use crate::Row;
 
 #[derive(Clone)]
 pub enum RusqliteParam {
@@ -53,36 +47,6 @@ impl From<&str> for RusqliteParam {
 impl From<Option<&str>> for RusqliteParam {
     fn from(value: Option<&str>) -> Self {
         Self::Value(value.map(|s| s.to_string()).into())
-    }
-}
-
-pub struct RusqliteRowBorrowed<'a, 'b> {
-    pub(super) row: &'a rusqlite::Row<'b>,
-}
-
-impl Row for RusqliteRowBorrowed<'_, '_> {
-    fn get_int(&mut self, idx: &str) -> i64 {
-        self.row.get(idx).unwrap()
-    }
-
-    fn get_opt_int(&mut self, idx: &str) -> Option<i64> {
-        self.row.get(idx).unwrap()
-    }
-
-    fn get_text(&mut self, idx: &str) -> String {
-        self.row.get(idx).unwrap()
-    }
-
-    fn get_opt_text(&mut self, idx: &str) -> Option<String> {
-        self.row.get(idx).unwrap()
-    }
-
-    fn get_blob(&mut self, idx: &str) -> Vec<u8> {
-        self.row.get(idx).unwrap()
-    }
-
-    fn get_blob_array<const N: usize>(&mut self, idx: &str) -> [u8; N] {
-        self.row.get(idx).unwrap()
     }
 }
